make time range consistent with district data
For the state data, we now calculate cases & tests in the 7 days prior to the current date instead of including the current date. This makes the values more consistent with the district data from MOHFW, and also more stable (.e. less likely to fluctuate as the data is being updated for the data).